Split event mutation controllers per endpoint
